+++ tests/test-lang-definition
@@ -7,7 +7,13 @@ p = PromptClient(pulsar_host="pulsar://localhost:6650")
 
 chunk = """I noticed a cat in my garden.  It is a four-legged animal
 which is a mammal and can be tame or wild.  I wonder if it will be friends
-with me.  I think the cat's name is Fred and it has 4 legs"""
+with me.  I think the cat's name is Fred and it has 4 legs.
+
+A cat is a small mammal.
+
+A grapefruit is a citrus fruit.
+
+"""
 
 resp = p.request_definitions(
     chunk=chunk,
