+++ trustgraph-flow/trustgraph/model/prompt/template/service.py
@@ -155,17 +155,6 @@ class Processor(ConsumerProducer):
             config = prompt_configuration,
         )
 
-    def parse_json(self, text):
-        json_match = re.search(r'```(?:json)?(.*?)```', text, re.DOTALL)
-    
-        if json_match:
-            json_str = json_match.group(1).strip()
-        else:
-            # If no delimiters, assume the entire output is JSON
-            json_str = text.strip()
-
-        return json.loads(json_str)
-
     def handle(self, msg):
 
         v = msg.value()
